backend/requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
pymysql==1.1.1
alembic==1.13.1
cryptography==42.0.8

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# Validation
pydantic==2.7.4
pydantic-settings==2.3.4
email-validator==2.1.1

# HTTP / Async
httpx==0.27.0
aiofiles==23.2.1

# Rate limiting
slowapi==0.1.9

# File processing
python-magic==0.4.27
pillow==10.4.0
pypdf2==3.0.1
python-docx==1.1.2

# AI / ML
anthropic==0.28.1
openai==1.35.10
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2

# Reporting / PDF
reportlab==4.2.2
weasyprint==62.3
jinja2==3.1.4

# Email
fastapi-mail==1.4.1

# Caching
redis==5.0.6
hiredis==2.3.2

# Background tasks
celery==5.4.0

# Utils
python-dateutil==2.9.0
pytz==2024.1
Babel==2.15.0

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
pytest-cov==5.0.0
factory-boy==3.3.0

# Dev
python-dotenv==1.0.1
